fix: Small refactor on useDecision hook
