+++ src/experiments/index.jsx
@@ -1,0 +1,5 @@
+import ExperimentsProvider from './ExperimentsProvider';
+
+export * from './experimentHooks';
+
+export { ExperimentsProvider };
