--- src/optimizely/index.jsx
@@ -1,3 +1,0 @@
-import ExperimentsProvider from './ExperimentsProvider';
-
-export default ExperimentsProvider;
